requirements.txt
wavio
sounddevice
faiss-cpu
langchain
crewai
crewai-tools
openai-whisper
transformers
sentence-transformers
langchain-community
alpha_vantage
streamlit
fastapi
uvicorn
soundfile
python-multipart
langchain-openai
mangum
openai
ffmpeg